--- TwoHoSun/Data/Network/Foundation/ErrorResponse.swift
@@ -1,21 +1,0 @@
-//
-//  ErrorResponse.swift
-//  TwoHoSun
-//
-//  Created by 235 on 11/11/23.
-//
-
-import Foundation
-
-struct ErrorResponse: Decodable, Error {
-    let status: Int
-    let divisionCode: String
-    let message: String
-}
-
-struct AuthErrorResponse: Decodable {
-    let status: Int
-    let divisionCode: String
-    let message: String
-    let data: TokenObject
-}
