--- TwoHoSun/Model/Prev/LoginModel/NicknameValidation.swift
@@ -1,12 +1,0 @@
-//
-//  NicknameModel.swift
-//  TwoHoSun
-//
-//  Created by 관식 on 10/17/23.
-//
-
-import Foundation
-
-struct NicknameValidation: Codable {
-    var isExist: Bool
-}
