+++ TwoHoSun/Util/Types/SchoolType.swift
@@ -1,0 +1,13 @@
+//
+//  SchoolType.swift
+//  TwoHoSun
+//
+//  Created by 관식 on 10/23/23.
+//
+
+import Foundation
+
+enum SchoolType: String, Codable {
+    case middleSchool = "중학교"
+    case highSchool = "고등학교"
+}
