+++ TwoHoSun/Util/Types/TermAgreeType.swift
